[tools/color correction] better agreement with standards
The ranges of the sliders have been modified to directly
correspond to the CDL parameters, and the limits have
been increased to have more room for adjustments.
The offset slider has a x10 factor for better fine control.

Loading/saving of CDL files (in XML format) has been added to the UI