[gui] improved image closing
The message dialog that warns the user when the
image being closed contains unsaved data has been improved.
Now it provides an option to "cancel" the closing process,
and go back to the edited image.

Additionally, if the user decides to save the file
but then decides to cancel the save process by clicking
on the "Cancel" button of the file dialog,
the closing process is cancelled as if the user had
selected "Cancel" in the first message dialog.