[general] implemented default input for nodes
Each node has now a default input, which is selected depending on the pipeline layout and what the node accepts as input colorspace. Layer mask nodes can thus automatically get a grayscale version of the image as input, making the introduction of a clone layer not necessary anymore. The clone layer tool has been renamed to "channel selector".