[tools/raw developer] forced demosaicing at all previw zoom levels
Then RAW loader now always returns a full-resolution raw image, so that the raw developer always performs the demosaicing and other operations on raw pixels. The output is then down-scaled to the requested zoom level.
This fixes a mismatch of the previews at 1:1 and below, due to the different raw pixel processing used in the two cases.