[base] introduce reuse of nodes from previous pipelines in chain rebuild
When building a node at a given level, the code now checks if an equivalent node was already created at the same zoom level for another pipeline. In this case, the VipsImage objects associated with the already existing node are simply referenced. This allows a more efficient re-use of tile caches.