[tools/sharpening] further improvements to the enhanced USM method
The noise preservation is now implemented with an "inverted guided filter" that blurs only edges stronger than the threshold.
For this, a modified version of the guided filter has been implemented, which outputs the "a" parameter and the mean values instead of the final linear combination between the input and the mean. This allows the inversion of the guided filter behavior by replacing "a" with (1-a).
The call to the modified guided filter has been also integrated into the enhanced USM code for better performances.